Fix sign-compare and dangling-else warnings
.../src/GenEvent.cc:358:28: warning: comparison of integer expressions of different signedness: 'const int' and 'std::vector<std::shared_ptr<HepMC3::GenVertex> >::size_type' {aka 'long unsigned int'} [-Wsign-compare]

.../src/ReaderLHEF.cc:80:41: warning: comparison of integer expressions of different signedness: 'int' and 'size_t' {aka 'long unsigned int'} [-Wsign-compare]

.../src/ReaderLHEF.cc:145:85: warning: comparison of integer expressions of different signedness: 'int' and 'std::vector<std::shared_ptr<HepMC3::GenParticle> >::size_type' {aka 'long unsigned int'} [-Wsign-compare]

.../src/ReaderLHEF.cc:149:54: warning: suggest explicit braces to avoid ambiguous 'else' [-Wdangling-else]

+++ src/GenEvent.cc
@@ -355,7 +355,7 @@ void GenEvent::add_tree( const std::vector<GenParticlePtr> &parts ) {
     if ( m_rootvertex->id() != 0 ) {
         const int vx = -1 - m_rootvertex->id();
         const int rootid = m_rootvertex->id();
-        if ( vx >= 0 && vx < m_vertices.size() && m_vertices[vx] == m_rootvertex ) {
+        if ( vx >= 0 && vx < (int) m_vertices.size() && m_vertices[vx] == m_rootvertex ) {
             auto next = m_vertices.erase(m_vertices.begin() + vx);
             std::lock_guard<std::recursive_mutex> lock(m_lock_attributes);
             for(auto & vt1: m_attributes ) {
